moved class definitions, calling api also called database_init causing duplicates
